backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
requests==2.31.0
gunicorn==21.2.0

